Add author adding/choosing logic.
- Will provide list with existing authors
- Will select the correct one that is assigned to an app.
- Will create and de-dup newly entered author

+++ appengine/src/main/java/org/retrostore/data/app/AppStoreItem.java
@@ -140,7 +140,7 @@ public class AppStoreItem {
     public Set<ListingCategory> categories = new HashSet<>();
     public long firstPublishTime;
     public long lastUpdateTime;
-    public int authorId;
+    public long authorId;
     public String publisherEmail;
   }
